+++ CHANGES.md
@@ -17,6 +17,8 @@
   * Added missing instances for overflow-checking arithmetic of arbitrary
     sized signed and unsigned bitvectors.
 
+  * Improve SMTLib logic-detection code to use generics.
+
 ### Version 8.8, 2020-09-04
 
   * Reworked uninterpreted sorts. Added new function `mkUninterpretedSort` to make